Refactorized InMemory test class
Created an abstract class to extends

TODO: Move to ddd-test package
